server/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
requests==2.26.0
together==0.2.1
googlemaps==4.10.0
python-jose==3.3.0
python-dateutil==2.8.2
gunicorn==20.1.0 